AOC refactore - move java code to main/java
