+++ src/main/java/hr/bp/aoc/InputUtil.java
