+++ src/main/java/hr/bp/aoc/day1/CalorieCounting.java
