+++ src/main/java/hr/bp/aoc/day10/CathodeRayTube.java
