+++ src/main/java/hr/bp/aoc/day3/PrioritiesUtil.java
