+++ src/main/java/hr/bp/aoc/day4/PairsUtil.java
