+++ src/main/java/hr/bp/aoc/day5/CratesUtil.java
