+++ src/main/java/hr/bp/aoc/day5/SupplyStacks.java
