+++ src/main/java/hr/bp/aoc/day6/MarkerUtil.java
