+++ src/main/java/hr/bp/aoc/day6/TuningTrouble.java
