+++ src/main/java/hr/bp/aoc/day7/Directory.java
