+++ src/main/java/hr/bp/aoc/day7/DirectoryUtil.java
