+++ src/main/java/hr/bp/aoc/day7/NoSpaceLeftOnDevice.java
