+++ src/main/java/hr/bp/aoc/day8/ScoreUtil.java
