+++ src/main/java/hr/bp/aoc/day8/TreetopTreeHouse.java
