+++ src/main/java/hr/bp/aoc/day9/Position.java
