+++ src/main/java/hr/bp/aoc/day9/RopeBridge.java
